crypto: skcipher - Add SKCIPHER_REQUEST_ON_STACK
Add SKCIPHER_REQUEST_ON_STACK which places a generic skcipher request
on the stack.  If an operation on it cannot complete synchrously, it
will fail with EAGAIN.  The request should then be converted to a
dynamically allocated one using SKCIPHER_REQUEST_CLONE.  If memory
allocation fails the request will automatically switch to a fallback
that is synchronous.

If a stack request is given to skcipher_request_free it will simply
be zeroed.

Finally add skcipher_request_alloc_extra which gives the user extra
memory to use in conjunction with the request.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>